refactor: UserJooqRepository를 service 패키지로 이동하고 UserQueryService로 변경
