+++ src/main/java/com/codeit/team4/deokhugam/user/service/UserService.java
@@ -9,7 +9,6 @@ import com.codeit.team4.deokhugam.user.dto.UserResponse;
 import com.codeit.team4.deokhugam.user.dto.UserUpdateRequest;
 import com.codeit.team4.deokhugam.user.entity.User;
 import com.codeit.team4.deokhugam.user.mapper.UserMapper;
-import com.codeit.team4.deokhugam.user.repository.UserJooqRepository;
 import com.codeit.team4.deokhugam.user.repository.UserRepository;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
@@ -29,7 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final UserJooqRepository userJooqRepository;
+    private final UserQueryService userQueryService;
     private final UserMapper userMapper;
 
     public UserResponse getUser(UUID userId) {
@@ -122,7 +121,7 @@ public class UserService {
 
         Instant threshold = Instant.now().minus(1, ChronoUnit.DAYS);
 
-        int deletedCount = userJooqRepository.deleteExpiredUsers(threshold);
+        int deletedCount = userQueryService.deleteExpiredUsers(threshold);
 
         log.info("유저 물리 삭제 완료: count={}, threshold={}", deletedCount, threshold);
     }
